<WIP> Changing text colors in the different rock descriptions
In this version, I finished the first part of changing the colors of the keys.

For example, now when you tap "Sandstone", if you scroll down to where it says "Safe" along with the meaning behind that word, the text will now be a different color!

Because it's late, I didn't get to change the search_result.dart file yet to reflect these changes. Will happen tomorrow though!