<WIP?> Bug Fixed where settingsPage wasn't updated to show what units weatherPage used.
This fixed an issue I saw after the last version. If the user switched units -- let's say from F to C -- the content in the weatherPage and searchResultsPage would update, but not the settingsPage. If after the user came back to the settings page, they would still be Imperial units (even though they switched units). This update should fix that by making sure that settingsPage uses sharedPrefrences.