<WIP> Might Need to implement state management
Anytime the user clicks home, the weather is fecthed. However, I only want it to be fetched under these conditions:

Running the app (and it's not been in the background) for the first time.
It's been more than an hour since last fetch.
Units have been changed (like F to C or vice versa).