OpenWeatherMapAPIKey has been hidden!
I have made some changes to the application that allows the application to be more secure. Instead of hardcoding the OWM-API, I used FlutterSecureStorage to hide the keys. I'm not sure if it can still be cracked or opened as the plain text API key is in the main.dart file.

NOTE: GOOGLE MAPS API KEY HAS NOT BEEN SECURED YET. WILL BE IN NEXT VERSION UPDATE.