Updated screens to incorporate units for wind speed
This feature is fully done.

----------------------------
Todo: Fix states. What I mean is that if you keep navigating around, the pushState keeps adding screens to the stack. I think what's best is this: keep everything off the stack besides these two conditions -- 1) The user is on searchResult and wants to go back to searchScreen. 2) Check to see if weatherPage is on the stack. If it is, then go to that page with said state, otherwise create weatherPage and add to the stack. The goal for this is so that anytime the user clicks home, it doesn't always fetch API data and hopefully saves us some requests. Furthermore, weatherPage should only make API calls under these conditions: 1) weatherPage isn't on the stack (because the app has just been started or restarted). 2) Because the unitType has changed (the user switched from imperial to metric or vice versa). 3) It's been over an hour since the last call.
Todo: Investigate 36_to_72 hour and over_72_hour variables since I feel like they're always set to false?
Todo: Clean up code, add comments, make code readable using DartDoc.