UnitChanges don't work, states do work.
In this version, I changed the code so that anytime the user presses the home button in the page it doesn't auto get information from the API. Instead, it checks to see if there's internet (if not don't even try to fetch data), then see if it's a new app instance (not in background) (and if so, get weather), and checks to see if it's been over an hour since last update (if it has, get new weather info).

EVERYTHING WORKS EXCEPTIONAL EXCEPT: If the user goes to settings, changes units, and goes anywhere else on the page, the units get reset and reverted. For example, if the user changes from metric to imperial and goes to either home or search, upon returning to the settings, the units are still metric. Will work on it tomorrow.