This code is on this branch to be used once I gain access to the historicalweatherapi.
My student account did not give me access to the historical weather API when it should. I sent a request to the support team, but they've yet to respond. I'll merge this branch if I'm given access to it, otherwise I'll only have access to the last 24 hours of data.